refactor checksum and use it for path (file) sources
